# specs/001-headless-financial-api/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Tazco Financial API
  description: |
    Headless financial backend service for credit card management, user scoring,
    and payment simulation. All state changes are published to a message stream
    for downstream consumers.
  version: 1.0.0
  contact:
    name: Tazco Platform Team
servers:
  - url: https://{region}-{project}.cloudfunctions.net/api
    description: Firebase Functions (Production)
    variables:
      region:
        default: us-central1
      project:
        default: tazco-financial
  - url: http://localhost:5001/{project}/{region}/api
    description: Firebase Emulator (Development)
    variables:
      region:
        default: us-central1
      project:
        default: tazco-financial
...
security:
  - bearerAuth: []
paths:
...
  /v1/cards:
    get:
      tags: [Cards]
      summary: List user's cards
      description: Returns all cards owned by the authenticated user.
      operationId: listCards
      parameters:
        - name: type
          in: query
          description: Filter by card type
          schema:
            type: string
            enum: [credit-card]
        - name: status
          in: query
          description: Filter by status
          schema:
            type: string
            enum: [active, suspended, cancelled]
      responses:
        '200':
          description: List of cards
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardsListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/cards/{cardId}:
    get:
      tags: [Cards]
      summary: Get card details
      description: Returns detailed information about a specific card.
      operationId: getCard
      parameters:
        - $ref: '#/components/parameters/CardId'
      responses:
        '200':
          description: Card details
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/cards/requests:
    post:
      tags: [Cards]
      summary: Request a new card
      description: |
        Submit a credit card application. Decision based on user's score:
        - Score >= 700: Auto-approved with $10,000 limit
        - Score 500-699: Auto-approved with $5,000 limit
        - Score < 500: Pending admin review
      operationId: requestCard
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardRequestInput'
      responses:
        '201':
          description: Card request created (may be approved or pending)
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardRequestResponse'
        '200':
          description: Card request already processed (idempotent replay)
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardRequestResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Conflict (existing pending request or active card)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: CARD_REQUEST_EXISTS
                  message: You already have a pending card request
                  details:
                    existingRequestId: req_abc123
  /v1/cards/requests/{requestId}:
    get:
      tags: [Cards]
      summary: Get card request status
      description: Returns current status and decision (if any) for a card request.
      operationId: getCardRequest
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Card request details
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardRequestResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================================
  # TRANSACTIONS
  # ============================================================================
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase Authentication JWT token

  # ============================================================================
  # PARAMETERS
  # ============================================================================
  parameters:
    CardId:
      name: cardId
      in: path
      required: true
      description: Card identifier
      schema:
        type: string
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Client-generated unique key for request deduplication (valid 24h)
      schema:
        type: string
        maxLength: 64
...
  headers:
    X-Request-ID:
      description: Unique request identifier for tracing
      schema:
        type: string
        format: uuid

  # ============================================================================
  # RESPONSES
  # ============================================================================
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            token_expired:
              value:
                error:
                  code: TOKEN_EXPIRED
                  message: Authentication token has expired
            invalid_token:
              value:
                error:
                  code: INVALID_TOKEN
                  message: Authentication token is invalid
            account_disabled:
              value:
                error:
                  code: ACCOUNT_DISABLED
                  message: Account has been disabled
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  # ============================================================================
  # SCHEMAS
  # ============================================================================
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            details:
              type: object
              additionalProperties: true
              description: Additional error context
...
    CardSummary:
      type: object
      required: [cardId, type, status, limit, balance, availableCredit]
      properties:
        cardId:
          type: string
        type:
          type: string
        status:
          type: string
          enum: [active, suspended, cancelled]
        limit:
          type: number
          format: double
        balance:
          type: number
          format: double
        availableCredit:
          type: number
          format: double
        minimumPayment:
          type: number
          format: double
        nextDueDate:
          type: string
          format: date-time
        nearLimit:
          type: boolean
          description: True if utilization > 90%
...
    CardsListResponse:
      type: object
      required: [cards]
      properties:
        cards:
          type: array
          items:
            $ref: '#/components/schemas/CardSummary'
        suggestion:
          type: string
          description: Suggestion if no cards exist
    CardResponse:
      type: object
      required: [card]
      properties:
        card:
          $ref: '#/components/schemas/CardDetail'
    CardDetail:
      allOf:
        - $ref: '#/components/schemas/CardSummary'
        - type: object
          properties:
            createdAt:
              type: string
              format: date-time
            activatedAt:
              type: string
              format: date-time
            approvedBy:
              type: string
              enum: [auto, admin]
            scoreAtApproval:
              type: integer
    CardRequestInput:
      type: object
      required: [productId]
      properties:
        productId:
          type: string
          description: Product identifier (e.g., 'standard-credit-card')
    CardRequestResponse:
      type: object
      required: [request]
      properties:
        request:
          type: object
          required: [requestId, status, scoreAtRequest]
          properties:
            requestId:
              type: string
            status:
              type: string
              enum: [pending, approved, rejected]
            scoreAtRequest:
              type: integer
            decision:
              type: object
              properties:
                outcome:
                  type: string
                  enum: [approved, rejected]
                source:
                  type: string
                  enum: [auto, admin]
                approvedLimit:
                  type: number
                reason:
                  type: string
            card:
              $ref: '#/components/schemas/CardSummary'
              description: Included if approved

    # --------------------------------------------------------------------------
    # Transactions
    # --------------------------------------------------------------------------
...